fix(parallel-keyset): 5 bugs from the multi-agent branch bughunt
A 6-dimension adversarial bughunt over the branch diff (each finding verified at
its named site) surfaced 5 distinct defects (one found twice):

1. [HIGH] keyset_range PG DDL is int4, bound/read as i64 (src/state/mod.rs).
   The v19 PG migration declared `range_index INTEGER` / `done INTEGER`, but the
   state layer binds them as StateParam::I64 and reads via StateRow::i64, and
   rust-postgres is strictly typed (i64 <-> INT8 only). On a Postgres STATE backend
   (RIVET_STATE_URL), parallel keyset + chunk_checkpoint errored (WrongType) at OPEN
   in persist_keyset_ranges and panicked on resume in load_keyset_ranges — the
   feature was dead-on-arrival on PG state. SQLite (default, dynamically typed)
   coerced fine, so every SQLite-only unit test hid it. Fix: declare both BIGINT,
   matching every other integer column in PG_MIGRATIONS. New live round-trip test
   pg_keyset_range_round_trips_and_commits (RED on int4, green on BIGINT).

2. [MEDIUM] MySQL boundary sampling ran inline literals on an UNPINNED connection
   (src/source/mysql/mod.rs). The M1 sql_mode pin (strip NO_BACKSLASH_ESCAPES) lived
   only in MysqlSessionGuard applied in export(); sample_key_boundaries injects
   escaped floor/ceil literals via query_scalar on a fresh pool connection with the
   server-default sql_mode. Under NO_BACKSLASH_ESCAPES a backslash-bearing text
   anchor mis-parses, drifting the percentile split (lost fan-out) or, on an
   incremental re-run, duplicating rows across the boundary. Fix: apply the same
   MysqlSessionGuard to the query_scalar connection so sampler == worker.

3. [MEDIUM] check_post_run_invariants exempted resumes via `self.resumed`
   (adopted_prior_work), not the --resume invocation (src/pipeline/summary.rs). A
   chunked-checkpoint run that crashed BEFORE its first chunk committed adopts no
   prior work (resumed=false) yet still legitimately skips the drift gate, so the
   guard panicked a legitimate crash-recovery resume in debug/test builds. Fix:
   thread plan.resume as is_resume_run and exempt on `!resumed && !is_resume_run`.
   RED-proven unit test (is_resume_run=true passes; false still trips).

4. [LOW] Non-checkpoint parallel keyset wrote NO file_log rows (src/pipeline/keyset.rs).
   The merge loop passed state=None unconditionally, correct only for checkpoint runs
   (workers wrote file_log). A non-checkpoint parallel run left `rivet state files`
   empty while the byte-identical sequential run listed all N. No data loss (manifest
   complete). Fix: pass `if checkpoint { None } else { state }`.

5. [LOW] Golden catalog keyset_sparse_unsigned key_min/key_max were ~10^16 (BELOW
   i64::MAX), contradicting the fixture (10^19 + n*10^8, above i64::MAX) and the row's
   own 'ENTIRELY above i64::MAX' guard — three zeros dropped. Fix: correct to the
   20-digit bounds; rebuild golden_catalog.db.

The 1 refuted candidate (a stale "20" file-count in a golden verify.sh comment) was
cosmetic — the enforced value is 21.

+++ src/state/mod.rs
@@ -665,15 +665,20 @@ const PG_MIGRATIONS: &[(i64, &str)] = &[
          ALTER TABLE export_metrics ADD COLUMN server_context_json TEXT;",
     ),
     // v19: parallel-keyset crash-recovery ranges (see the SQLite v19 comment).
+    // range_index/done are BIGINT (not the SQLite INTEGER): the state layer binds
+    // them as StateParam::I64 and reads them via StateRow::i64, and rust-postgres
+    // is strictly typed (i64 <-> INT8 only) — an int4 column would reject the bind
+    // (WrongType at persist) and panic on the read (resume). Every other integer
+    // column in PG_MIGRATIONS is BIGINT for exactly this reason.
     (
         19,
         "CREATE TABLE IF NOT EXISTS keyset_range (
             export_name TEXT NOT NULL,
             run_id      TEXT NOT NULL,
-            range_index INTEGER NOT NULL,
+            range_index BIGINT NOT NULL,
             lo          TEXT,
             hi          TEXT,
-            done        INTEGER NOT NULL DEFAULT 0,
+            done        BIGINT NOT NULL DEFAULT 0,
             updated_at  TEXT NOT NULL,
             PRIMARY KEY (export_name, range_index)
         );",
